Move story utility files to subfolders
