+++ packages/block-editor/src/components/inserter/stories/index.js
@@ -3,7 +3,7 @@
  */
 import BlockLibrary from '../library';
 import BlockEditorProvider from '../../provider';
-import { patternCategories, patterns, reusableBlocks } from './fixtures';
+import { patternCategories, patterns, reusableBlocks } from './utils/fixtures';
 import Inserter from '../';
 
 export default { title: 'BlockEditor/Inserter' };
