+++ packages/components/src/navigation/stories/index.js
@@ -6,12 +6,12 @@ import NavigationBackButton from '../back-button';
 import NavigationGroup from '../group';
 import NavigationItem from '../item';
 import NavigationMenu from '../menu';
-import { DefaultStory } from './default';
-import { GroupStory } from './group';
-import { ControlledStateStory } from './controlled-state';
-import { SearchStory } from './search';
-import { MoreExamplesStory } from './more-examples';
-import { HideIfEmptyStory } from './hide-if-empty';
+import { DefaultStory } from './utils/default';
+import { GroupStory } from './utils/group';
+import { ControlledStateStory } from './utils/controlled-state';
+import { SearchStory } from './utils/search';
+import { MoreExamplesStory } from './utils/more-examples';
+import { HideIfEmptyStory } from './utils/hide-if-empty';
 import './style.css';
 
 export default {
