+++ packages/components/src/navigation/stories/utils/default.js
@@ -6,9 +6,9 @@ import { useState } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import Navigation from '..';
-import NavigationItem from '../item';
-import NavigationMenu from '../menu';
+import Navigation from '../..';
+import NavigationItem from '../../item';
+import NavigationMenu from '../../menu';
 
 export function DefaultStory() {
 	const [ activeItem, setActiveItem ] = useState( 'item-1' );
